store/wizard: extract system types into dedicated module
Move SystemSlice and related user/group types out of the top-level
wizard types file into wizard/system/ so they live closer to the
slice that owns them.
