+++ src/store/slices/wizard/index.ts
@@ -10,3 +10,4 @@ export * from './details';
 export * from './filesystem';
 export * from './output';
 export * from './registration';
+export * from './system';
